add some styles and an image header
