+++ src/components/Pages/ResidentialParkingPermit/_ResidentParkingPermit.scss
@@ -1,0 +1,9 @@
+.coa-knack-rrp {
+  text-align: center;
+  .big-button-container {
+    margin: auto;
+    padding: 20px 10px;
+    font-size: 32px;
+    line-height: 36px;
+  }
+}
