+++ src/css/coa.scss
@@ -26,6 +26,7 @@
 @import 'components/HeroHome/HeroHome';
 @import 'components/HtmlFromAdmin/HtmlFromAdmin';
 @import 'components/ListLink/ListLink';
+@import 'components/Pages/ResidentialParkingPermit/ResidentParkingPermit';
 @import 'components/PageBanner/PageBanner';
 @import 'components/PageBreadcrumbs/PageBreadcrumbs';
 @import 'components/PageHeader/PageHeader';
